@@ -1,6 +1,6 @@
 module github.com/o-mago/spotify-status
 
-go 1.18
+go 1.21
 
 require (
 	github.com/google/uuid v1.3.0
